add article type
